refactorizacion: customers y transactions coherentes
